Move remaining php81 tests into the php81 test suite
