+++ tests/php81/EnumParameterTest.php
@@ -1,46 +1,35 @@
 <?php
 
-namespace Builders;
+namespace php81;
 
+use ClassWithEnumDependency;
 use lucatume\DI52\Builders\Parameter;
 use PHPUnit\Framework\TestCase;
 use ReflectionMethod;
-use UnionTypeClass;
-use UnionTypePromotedClass;
+use TestBackedEnum;
+use UnionTypeEnumClass;
 
-class UnionTypeParameterTest extends TestCase {
+class EnumParameterTest extends TestCase
+{
 
     /**
      * @beforeClass
      */
     public static function before_all()
     {
-        if (PHP_VERSION_ID < 80000) {
-            return;
-        }
-
-        require_once __DIR__.'/parameter-test-union-type-classes.php';
+        require_once __DIR__.'/data/parameter-test-enum-classes.php';
     }
 
-    /**
-     * @before
-     */
-    public function before_each()
-    {
-        if (PHP_VERSION_ID < 80000) {
-            $this->markTestSkipped();
-        }
-    }
-    public function test_it_should_detect_union_types()
+    public function test_it_should_detect_enum_types()
     {
         $expectedData = [
             [
-                'type' => 'union',
+                'type' => TestBackedEnum::class,
                 'isOptional' => false,
                 'defaultValue' => null,
             ]
         ];
-        $reflectionConstructor = new ReflectionMethod(UnionTypeClass::class, '__construct');
+        $reflectionConstructor = new ReflectionMethod(ClassWithEnumDependency::class, '__construct');
         foreach ($reflectionConstructor->getParameters() as $i => $p) {
             $parameter = new Parameter($i, $p);
             $data = $parameter->getData();
@@ -51,7 +40,7 @@ class UnionTypeParameterTest extends TestCase {
         }
     }
 
-    public function test_it_should_detect_union_types_with_constructor_promotion()
+    public function test_it_should_detect_union_types_with_enums()
     {
         $expectedData = [
             [
@@ -60,7 +49,7 @@ class UnionTypeParameterTest extends TestCase {
                 'defaultValue' => null,
             ]
         ];
-        $reflectionConstructor = new ReflectionMethod(UnionTypePromotedClass::class, '__construct');
+        $reflectionConstructor = new ReflectionMethod(UnionTypeEnumClass::class, '__construct');
         foreach ($reflectionConstructor->getParameters() as $i => $p) {
             $parameter = new Parameter($i, $p);
             $data = $parameter->getData();
