+++ tests/php81/data/parameter-test-enum-classes.php
